refactor: structure of project
