+++ src/components/choose-area/choose-area.js
