+++ src/components/choose-area/choose-area.scss
