+++ src/components/dialogue-area/dialogue-area.scss
