+++ src/components/enter-name/enter-name.scss
