+++ src/components/game-area/game-area.js
