+++ src/components/modal/modal.scss
