+++ src/components/scoreboard/scoreboard.js
