+++ src/components/scoreboard/scoreboard.scss
