+++ src/components/select-setting/select-setting.js
