+++ src/components/select-setting/select-setting.scss
