+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.scss';
 import './normolize.min.scss';
-import App from './App/App';
+import App from './components/App/App';
 
 ReactDOM.render(
   <React.StrictMode>
